using mp instead of dask as default.
reasoning: the benefit of dask is the profiler and the dash board. Both only have a use, when the user actually intends to look at them. And if a user wants to look at them, they should turn on the CLI flag. User who don't care about this feature are better served with mp. Thus the default is now mp